pbs-middleware-common/src/main/java/com/pbs/middleware/common/validations/EmptyStringValidator.java
package com.pbs.middleware.common.validations;

import javax.validation.ConstraintValidator;
import javax.validation.ConstraintValidatorContext;
import lombok.RequiredArgsConstructor;

@RequiredArgsConstructor
public class EmptyStringValidator implements ConstraintValidator<NotEmptyString, String> {

    @Override
    public boolean isValid(String value, ConstraintValidatorContext context) {
        if (value == null) {
            return true;
        }
        return !value.isBlank();
    }
}
